refactor(@angular-devkit/build-angular): move esbuild compiler plugin in angular directory
Minor directory restructure to better organize the Angular compiler plugin code files.
The remainder of the Angular specific files for the Angular esbuild compiler plugin are now
within the `angular` subdirectory of the esbuild-based browser application builder.
