+++ packages/angular_devkit/build_angular/src/builders/browser-esbuild/angular/compiler-plugin.ts
@@ -20,19 +20,19 @@ import { platform } from 'node:os';
 import * as path from 'node:path';
 import { pathToFileURL } from 'node:url';
 import ts from 'typescript';
-import { maxWorkers } from '../../utils/environment-options';
-import { JitCompilation } from './angular/jit-compilation';
-import { setupJitPluginCallbacks } from './angular/jit-plugin-callbacks';
-import { AngularCompilation, FileEmitter } from './angular-compilation';
-import { AngularHostOptions } from './angular-host';
-import { JavaScriptTransformer } from './javascript-transformer';
+import { maxWorkers } from '../../../utils/environment-options';
+import { JavaScriptTransformer } from '../javascript-transformer';
 import {
   logCumulativeDurations,
   profileAsync,
   profileSync,
   resetCumulativeDurations,
-} from './profiling';
-import { BundleStylesheetOptions, bundleComponentStylesheet } from './stylesheets';
+} from '../profiling';
+import { BundleStylesheetOptions, bundleComponentStylesheet } from '../stylesheets';
+import { AngularHostOptions } from './angular-host';
+import { AotCompilation, FileEmitter } from './aot-compilation';
+import { JitCompilation } from './jit-compilation';
+import { setupJitPluginCallbacks } from './jit-plugin-callbacks';
 
 /**
  * Converts TypeScript Diagnostic related information into an esbuild compatible note object.
@@ -166,7 +166,7 @@ export function createCompilerPlugin(
       const javascriptTransformer = new JavaScriptTransformer(pluginOptions, maxWorkers);
 
       const { GLOBAL_DEFS_FOR_TERSER_WITH_AOT, readConfiguration } =
-        await AngularCompilation.loadCompilerCli();
+        await AotCompilation.loadCompilerCli();
 
       // Setup defines based on the values provided by the Angular compiler-cli
       build.initialOptions.define ??= {};
@@ -239,7 +239,7 @@ export function createCompilerPlugin(
 
       let stylesheetMetafiles: Metafile[];
 
-      let compilation: AngularCompilation | undefined;
+      let compilation: AotCompilation | undefined;
 
       build.onStart(async () => {
         const result: OnStartResult = {
@@ -291,7 +291,7 @@ export function createCompilerPlugin(
         if (pluginOptions.jit) {
           compilation ??= new JitCompilation();
         } else {
-          compilation ??= new AngularCompilation();
+          compilation ??= new AotCompilation();
         }
 
         // Initialize the Angular compilation for the current build.
