+++ packages/angular_devkit/build_angular/src/builders/browser-esbuild/index.ts
@@ -22,8 +22,8 @@ import { augmentAppWithServiceWorkerEsbuild } from '../../utils/service-worker';
 import { Spinner } from '../../utils/spinner';
 import { getSupportedBrowsers } from '../../utils/supported-browsers';
 import { BundleStats, generateBuildStatsTable } from '../../webpack/utils/stats';
+import { SourceFileCache, createCompilerPlugin } from './angular/compiler-plugin';
 import { checkCommonJSModules } from './commonjs-checker';
-import { SourceFileCache, createCompilerPlugin } from './compiler-plugin';
 import { BundlerContext, logMessages } from './esbuild';
 import { logExperimentalWarnings } from './experimental-warnings';
 import { createGlobalScriptsBundleOptions } from './global-scripts';
